Latest Google Shopping Crawl Version 2-4-2018
This version of the code reads from a .csv file containing product names and diststocks, then crawls the associated Google Shopping page for each diststock (aka part number). Then, for listings on the primary page that show competitor name as "from #X stores" it follows the associated link a level deeper and pulls in competitor listings on the detail page - including shipping and tax price information.

Data collected is then written to a .csv file with the name of today's date.